Re-organize region behaviors
