+++ module/data/_module.mjs
@@ -1,0 +1,3 @@
+export * as regionBehavior from "./region-behaviors/_module.mjs";
+
+export { default as TerrainData4e } from "./terrain-data.mjs";
