+++ module/data/region-behaviors/_module.mjs
@@ -1,0 +1,3 @@
+export { default as ApplyActiveEffect4eRegionBehaviorType } from "./apply-active-effect.mjs";
+export { default as DamagingRegionRegionBehaviorType } from "./damaging-region.mjs";
+export { default as DifficultTerrainRegionBehaviorType } from "./difficult-terrain.mjs";
