+++ module/data/region-behaviors/difficult-terrain.mjs
@@ -3,14 +3,13 @@
 const { BooleanField, NumberField, SetField, StringField } = foundry.data.fields;
 
 /**
- * @typedef DifficultTerrainRegionBehaviorSystemData
- * @property {boolean} magical                  This difficult terrain is caused by magic.
- * @property {Set<string>} types                Types of difficult terrain represented.
- * @property {Set<number>} ignoredDispositions  Token dispositions that won't be affected by this difficult terrain.
+ * @import { DifficultTerrainRegionBehaviorSystemData } from "./_types.mjs";
  */
 
 /**
  * The data model for a region behavior that represents an area of difficult terrain.
+ * * @extends {foundry.data.regionBehaviors.RegionBehaviorType<DifficultTerrainRegionBehaviorSystemData>}
+ * @mixes DifficultTerrainRegionBehaviorSystemData
  */
 export default class DifficultTerrainRegionBehaviorType extends foundry.data.regionBehaviors.RegionBehaviorType {
